@@ -1,5 +1,5 @@
 plugins {
-    id 'org.springframework.boot' version '2.7.5'
+    id 'org.springframework.boot' version '3.0.0'
     id 'io.spring.dependency-management' version '1.1.0'
     id 'java'
     id 'jacoco'
@@ -7,8 +7,8 @@ plugins {
 }
 
 group = 'crawler'
-version = '0.2.0'
-sourceCompatibility = '11'
+version = '1.0.0'
+sourceCompatibility = '17'
 
 configurations {
     developmentOnly
@@ -31,7 +31,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-webflux'
     implementation 'org.apache.lucene:lucene-analyzers-kuromoji:8.11.1'
-    implementation 'org.hibernate.search:hibernate-search-mapper-orm:6.1.7.Final'
+    implementation 'org.hibernate.search:hibernate-search-mapper-orm-orm6:6.1.7.Final'
     implementation 'org.hibernate.search:hibernate-search-backend-lucene:6.1.7.Final'
     compileOnly 'io.jsonwebtoken:jjwt-api:0.11.5'
     compileOnly 'org.projectlombok:lombok'
